+++ activestorage/lib/active_storage/transformers/image_processing_transformer.rb
@@ -1,6 +1,13 @@
 # frozen_string_literal: true
 
-require "image_processing"
+begin
+  require "image_processing"
+rescue LoadError
+  raise LoadError, <<~ERROR.squish
+    Generating image variants require the image_processing gem.
+    Please add `gem 'image_processing', '~> 1.2'` to your Gemfile.
+  ERROR
+end
 
 module ActiveStorage
   module Transformers