+++ activestorage/test/models/variant_test.rb
@@ -43,14 +43,10 @@ class ActiveStorage::VariantTest < ActiveSupport::TestCase
   end
 
   test "monochrome with default variant_processor" do
-    ActiveStorage.variant_processor = nil
-
     blob = create_file_blob(filename: "racecar.jpg")
     variant = blob.variant(monochrome: true).processed
     image = read_image(variant)
     assert_match(/Gray/, image.colorspace)
-  ensure
-    ActiveStorage.variant_processor = :mini_magick
   end
 
   test "disabled variation of JPEG blob" do
@@ -81,7 +77,6 @@ class ActiveStorage::VariantTest < ActiveSupport::TestCase
   end
 
   test "disabled variation using :combine_options" do
-    ActiveStorage.variant_processor = nil
     blob = create_file_blob(filename: "racecar.jpg")
     variant = ActiveSupport::Deprecation.silence do
       blob.variant(combine_options: {
@@ -95,12 +90,9 @@ class ActiveStorage::VariantTest < ActiveSupport::TestCase
     assert_equal 100, image.width
     assert_equal 100, image.height
     assert_match(/RGB/, image.colorspace)
-  ensure
-    ActiveStorage.variant_processor = :mini_magick
   end
 
   test "center-weighted crop of JPEG blob using :combine_options" do
-    ActiveStorage.variant_processor = nil
     blob = create_file_blob(filename: "racecar.jpg")
     variant = ActiveSupport::Deprecation.silence do
       blob.variant(combine_options: {
@@ -114,8 +106,6 @@ class ActiveStorage::VariantTest < ActiveSupport::TestCase
     image = read_image(variant)
     assert_equal 100, image.width
     assert_equal 100, image.height
-  ensure
-    ActiveStorage.variant_processor = :mini_magick
   end
 
   test "center-weighted crop of JPEG blob using :resize_to_fill" do